These tools assist in the assembly of NENA data.
DataCheck1 has several functions.  Change the target geodatasets in the
config.py file for datacheck 1, the target variables are at the
beginning of the Append_TEST.py file. 

A few of them can be ignored in favor of using the data reviewer
toolbar.  These tools check for basic geometry use and adherence to the
NG911 Template geodatabase for use in a statewide geodatabase. I a
geodatabase is not compliant to the template, new scripts will be
necessary to transfer data to and from a different schema.   

The LRSIfy adds some KDOT codes and a experimental method for creation
of nearly-unique, identifiable keys for LRS.  The key is XXX (county
number 001-105) + YYY  (City Number 010-620) + Z(coded LRS Prefix - I,
U, K, C, L)    A (coded street prefix/prefix dir) + B (Soundex of Named
Road, or padded street number)+C (aplha/numeric coded Road Type based on
most common values) + D(coded suffix direction) + U(Unique reservation)

The submitted data will be prepared for re-projection into the Kansas
Lambert Conformal Statewide Projection.  

Once the data is prepared by the DataCheck it can be loaded into a
statewide geodatabase by changing the target info at the top of the
script and running the append function, loading the entire contents of
the source geodatabase.    

!!!Be sure to change the functions you want to run at the bottom of the
script.!!!

!!!Be sure to use the DataCheck1 LRSify function prior to appending, or
it will not work!!!

Enter the Coded steward ID to delete the entire dataset from the
statewide dataset.